Added IsDone to the consumer and some more
